+++ src/modules/tools/indexTools.js
@@ -16,8 +16,10 @@ import FileImport from "./fileImport/store/indexFileImport";
 import GeneralFilter from "./generalFilter/store/indexGeneralFilter";
 import Gfi from "./gfi/store/indexGfi";
 import LayerClusterToggler from "./layerClusterToggler/store/indexLayerClusterToggler";
+import LayerSlider from "./layerSlider/store/indexLayerSlider";
 import Measure from "./measure/store/indexMeasure";
 import Print from "./print/store/indexPrint";
+import Routing from "./routing/store/indexRouting";
 import SaveSelection from "./saveSelection/store/indexSaveSelection";
 import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
 import SearchByCoord from "./searchByCoord/store/indexSearchByCoord";
@@ -25,7 +27,6 @@ import SelectFeatures from "./selectFeatures/store/indexSelectFeatures";
 import StyleVT from "./styleVT/store/indexStyleVT";
 import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
 import WfsSearch from "./wfsSearch/store/indexWfsSearch";
-import Routing from "./routing/store/indexRouting";
 
 /**
  * This is here to test app-store/utils/composeModules.
@@ -43,17 +44,18 @@ export default {
         FileImport,
         GeneralFilter,
         Gfi,
+        LayerClusterToggler,
+        LayerSlider,
         Measure,
         Print,
+        Routing,
         SaveSelection,
         ScaleSwitcher,
         SearchByCoord,
         SelectFeatures,
         StyleVT,
         SupplyCoord,
-        WfsSearch,
-        Routing,
-        LayerClusterToggler
+        WfsSearch
     },
     state,
     getters,