+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - added parameter "rotation" to wfs features with iconPointStyle, if the parameter is not given it will fall back to standard alignment.
 - The library "svg-url-loader" was added to package.json devDependencies.
 - Added the new Tool `Routing`.
+- Searchbar: An option "sortByName" in gdi to config if the rearching results from elastic sorted alphanumerically or not.
+- ParametricURL: An alert has been added which displays the new notation of a parameterized call.
+- Handling of polygons with holes to WKT geometry parsing.
 
 ### Changed
 - The following NPM packages are updated:
@@ -28,6 +31,8 @@
 - WMSLayer and GroupLayer are refactored. They are no longer Backbone-models. WMSLayer uses the masterportalAPI's wms layer on creation.
 - 2D-map is removed from vuex store. Maps are now stored in a collection. Creation of 2D-map and 3D-map use masterportalAPI's abstraction layer.
 - Print formats only contain working formats now.
+- QuickHelp: moved to Vue and refactored, can now be manipulated for search, tree and routing, new QuickHelp windows can be configured. See new quickHelp.md documentation for more details.
+- Migrated the layerSlider tool from Backbone.js to Vue.js.
 
 ### Deprecated
 
@@ -36,12 +41,18 @@
 - The library "olcs" was removed from the package.json.
 
 ### Fixed
-- Print tool: fixed wrong order of features in created print-map.
-- adding a File with other coordinate system may work now if the coordinate system in the JSON is EPSG 25832/4326 or can successfully be mapped to EPSG
+- Tool Print:
+    - Fixed wrong order of features in created print-map.
+    - If tool was activated by url parameter, print-mask is now visible and scales are selectable
+- Adding a File with other coordinate system may work now if the coordinate system in the JSON is EPSG 25832/4326 or can successfully be mapped to EPSG
 - Issue #654: WFS Layers didn't get displayed as group layers
 - Light-tree: Layers that are only selectable in certain zoom levels are now also grayed out directly after startup.
-- Print tool: if tool was activated by url parameter, print-mask is now visible and scales are selectable
-- Sidebar: if a tool was activated by url parameter and user opens second tool, the first tool is closed now
+- Tool wfsSearch: Zoom to a hit in the result list works now again
+- Sidebar: If a tool was activated by url parameter and user opens second tool, the first tool is closed now
+- Searchbar: Searches that deliver only a single search result can now be selected with `Enter`.
+- Handling of Multipolygons in WKT geometry parsing
+- GFI: in the iframe the content is now displayed again when using the `desktopType` `attached`.
+- GFI for theme datatable is displayed. Resolution at wms layer is set correctly.
 
 
 ---